Prefer unique constraints over indexes on SQL Server
Unique constraint and index do not differ in SQL Server today. The
problem is if we create an index and try to drop a constraint, the
drop fails.

Currently if the unique is added when the table is created, we add
constraint. If the table is altered with a unique constraint we create
an index instead. Dropping this does not work.

The PR makes altering the table with an `@unique` to use constraints
instead of an index.

Closes: https://github.com/prisma/prisma/issues/14051